chatonyourdata/requirements.txt
openai
streamlit
azure-storage-blob
streamlit-option-menu
python-dotenv